% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration}
\alias{calibration}
\title{Paper height calibration}
\arguments{
\item{lineus}{plotter (R6-Class)}

\item{x_limits}{minimum and maximum value of x-coordinates}

\item{y_limits}{minimum and maximum value of y-coordinates}

\item{z_corners}{2x2 matrix of calibration results (use phone app)}

\item{x}{x coordinate to retrieve z value for}

\item{y}{y coordinate to retrieve z value for}

\item{matrix}{matrix of interpolated z-coordinates}
}
\description{
This family of functions deals with the fact that the paper plane is not at \code{z=0}
}
