% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{operations}
\alias{operations}
\alias{up}
\alias{down}
\alias{move}
\alias{fly}
\alias{crawl}
\alias{draw}
\title{Operations}
\usage{
up(lineus)

down(lineus, x = NULL, y = NULL, z = NULL, canvas = NULL)

move(lineus, x = NULL, y = NULL, z = NULL)

fly(lineus, x, y)

crawl(lineus, x, y, canvas = NULL)

draw(lineus, from, to, cur = FALSE, canvas = NULL)
}
\arguments{
\item{lineus}{plotter (R6-Class)}

\item{x}{new x coordinate to move to}

\item{y}{new y coordinate to move to}

\item{z}{new z coordinate to move to}

\item{canvas}{Matrix with interpolated x coordinates, usually output of \code{\link[rlineus]{interpolate_z_matrix}}}

\item{cur}{start from current position?}
}
\value{
nothing
}
\description{
Functions to move the pen around
}
