% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{plotter_callback}
\alias{plotter_callback}
\title{callback function for plotting}
\usage{
plotter_callback(device_call, args, state)
}
\arguments{
\item{device_call}{call}

\item{args}{arguments}

\item{state}{state}
}
\description{
callback function for plotting
}
