% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{Test your plotter functionality}
\usage{
hello_world(what = c("boxplot", "polyline"))
}
\arguments{
\item{what}{string, either boxplot or polyline}
}
\value{

}
\description{
Try plotting a simple boxplot or sinus curve to check the plotter.
}
